Clean up Makefile and #includes
Cleaned up the Makefile to refer only to generic directory locations
(e.g. /usr/local/include instead of /home/jeffc/...) for AO headers.

Include hiredis.h as hiredis/hiredis.h to avoid having to pass
-I/usr/include/hiredis to the compiler.

Also change "-lserendip6" to "serendip6.so" for the GPU test program.
serendip6.so is not a regular shared object (it is a hashpipe plug-in)
and its name does not start with "lib..." so it cannot be linked in
using "-lserendip6".  We should probably refactor the code so that the
test program does not need to link to the hashpipe plugin.