Write scram field IF2SYNHZ to the FITS file.
I also commented out the writing of scram
field IF1IFFRQ to status shared memory.
We don't appear to use it for anyting and
it seems to always be zero during operations.
But I otherwise left it in the code, just
in case.

This "noting but not using" IF1IFFRQ seems
to be true in the seti@home code as well.