Fix reconfig bugs when num coarse channels changes.
The code specifies a hard coded maximum number of coarse channels.
At run time, this can be changed to a lesser value. When this occurs,
there are unused portions of the input data buffer.  There was a
difference between the net thread and gpu thread as to where these
unused portions were. This commit brings the gpu thread into agreement
with the net thread.