Clean up file change logic.
Upon change to receiver, number of coarse channels,
or RUNALWYS, the current file (if open) is immediately
closed and a flag is set to open a new file.

Among other things, this allows a file to be closed and
transfered to UCB when we go from some receiver to no
receiver, rather than waiting for a receiver to be in focus
again.