Major bug fix: we now analyze all coarse channels.
We had a bug that caused us to not analyze half of the
coarse channels that are received from the FPGAs.

I had to double the size of a number of CUDA vectors and
this put us over the amount of memory on the GPU.  So I went
to an on-demand scheme of allocating and de-allocating GPU
memory such that not all of the vectors are allocated at once.

Bumped the software version 0.3.0.