Rework missing packets diagnostics.
Changes the missing packets metrics in status shared memory.
We now see a total percentage missed and a percentage missed
per beam.

In this commit is also code to raise the scheduling priority of the thread
to maximum priority.  This is currently commented out.  To use this
feature, hashpipe must be run as root.