More on-demand GPU memory allocation.
When we went to the max number of coarse channels, we
once again ran out of memory.  So now we dynamically allocate
the big time series vector as well.

With the current GPU (GeForce GTX 780 Ti), our max GPU
meory usage is now 2988MB out of a total 3071MB on the
GPU.

There was a strange behavior with this change that I cannot
explain.  When we allocate the time series vector ahead of
the FFT vectors, the the address of the FFT out vector changes
with each re-allocation (the FFT in vector does not change
addresses in this situation).  When I allocate the time series
vector after those for the FFT, the FFT out (and in) vectors
have the same address over (apparently) all allocations.

Why is this?

Also, cuFFT throws a fatal exception when the FFT out
address changes, even though the new address is passed
to cuFFT.

Why it this?