Add and check scram field sigSrc.
The sigSrc field in the IF2 scram packet indicates the source
of the signal, with:

    0=gregorian, 1=carriage house, 2=noise

It is important to check this because, specifically with the
327MHz receiver, AO may have switched to the CH during
a system temperature check for all receivers.  During this
period, the LOs and turret position may still indicate that
the 327 is in focus (as indeed it is over in the gregorian dome).

If sigSrc != 0, then we do not acquire 327MHz data.

There may well be a subsequent change to do this check during
ALFA  observations.   I have a query in to AO about this.