Checkpoint of code changes for s6 at GBT.
The main difference between AO and GBT versions
is that we organize data by beam in the case of the
former and by sub-spectra in the case of the latter.

Note that this includes changes to s6_pktsock_thread.cpp
but not s6_net_thread.cpp.  s6_net_thread.cpp should
receive analogous changes at some point.