Log RMS's to status, do not output DC hits.
Course channel RMS's for both pols are now placed
into status shared memory for the coarse channel
of choice (also in status shared memory).

We no longer output hits in fine channel zero
(the DC bin).  This was done for the GBT deployment.
We may want to wrap this in an #ifdef SOURCE_DIBAS.