CLEO heartbeat
This implements the client side of the cleo heartbeat.

Once per minute (as of this commit), the cleo server
inserts a heartbeat into the data stream in the form
of key=Thump, value=Thump, along with the current
MJD.  While all other items in the data stream are
inserted only on a delta (and are thus unpredictable),
the heartbeat has a regular cadence.  Thus we can
detect when the data stream has gone down.  This
is treated as an error and serendip6 exits.

This supersedes (in the real time code) the cleo hello
method of problem detection.

I also slipped in a couple of unrelated comments.