Idle, rather than exit, on redis error.
Rather than exiting, s6 will now idle data output upon any sort
of redis error.  This could be an operational error like a failure
to connect to the redis server or a data error like detecting stale
meta data.

We found that such errors are typically transient.  When a clearing
of the error condition is detected, s6 de-idles and data output resumes.

S6HPCVER bumped to 0.7.3.

There are some dangling mods to coarse channel code in this commit
that should not cause a problem.  These will be resolved in a future
commit.