FAST specific code with temporary refactoring.
In order for the FAST development code to not get
too far behind either the "canonical" version at github
or that which is deployed at AO and the GBT, I sequestered
the FAST version of spectroscopy() as spectroscopy_one_pol()
and renamed the AO/GBT version as spectroscopy_two_pols().
The latter is an unchanged rename of the original spectroscopy().

As per the above, the main difference in the FAST version is that
it sends one pol at a time to the GPU.  This is because we are
dealing with very long FFTs with FAST.

There are a few other #ifdef's in this commit to sequester small
bits of FAST development code, such as slightly different GPU
dynamic memory allocation.

At some point, the one_pol / two_pols logic should be combined.
Otherwise, there is too much redundant code.