Clear/swap memory deallocation and other small changes.
Added a clear and swap option for GPU memory de-allocation in the FAST version of spectroscopy().
This is enabled via a #define. This may be a temporary change.

Bumped the input buffer size to hold 512*1024*1024 real time samples for FAST.

Added gpu_fini() for GPU cleanup.  Currently just calls cudaProfilerStop().

Added code to print device properties for the GPUs in the system.  Enabled via a #define.