Deleted commented-out input buff destination code.
Deleted a number if code lines dealing with the calculation
and debugging of the placement within the input buffer of
packet data.

All of these lines were commented out and so this commit
should have no functional effect.

The deletes were done to improve the readability of the code.

That all said, if there is a sudden problem with input data
placement, this would e a good place to start diagnosing the
problem!