Use the number of time samples in the calculation of Nm.
Use the number of time samples (N_TIME_SAMPLES) rather that the number
of fine channels (N_FINE_CHANNELS).  Fine channels should really refer to the
output of the FFT while time samples refers to the input voltage samples.

These two constants differ in the case of FAST but are equal in earlier s6
implementations.